audio: WASAPI shared-mode 32-bit float output (v1.1)
Render audio over WASAPI shared mode in 32-bit float, the native format
of the Windows mixer, instead of legacy 16-bit waveOut. The synthesis
engine is unchanged: all sample math stays in double precision and is now
emitted as float, so the signal no longer round-trips through a 16-bit
quantization step before reaching the mixer (the "headroom" win). The
event-driven render thread also tightens latency from parameter changes
to audible output.

The audio.h contract is unchanged, so main.c/gui.c/presets are untouched.
waveOut is kept as an automatic fallback (with a float->int16 conversion)
if WASAPI initialization fails, so playback always works.

- synth.{h,c}: fill buffers as float [-1,1] instead of int16
- audio.c: WASAPI render thread behind the existing API + waveOut fallback
- tests: update synth tests to float (24 tests pass)
- Makefile: link -lole32 for COM
- gui: window title -> "Tone Generator 1.1"
- docs: describe the new audio path in README.md / ReadMe.txt
- releases: add ToneGen-1.1-portable.zip

+++ src/synth.h
@@ -33,10 +33,11 @@ typedef struct {
     double mod_hz;
 } SynthFrameParams;
 
-/* Fills `out` with `frames` stereo (L,R) interleaved int16 samples.
- * `out` must point to space for at least 2*frames int16_t values.
+/* Fills `out` with `frames` stereo (L,R) interleaved float samples in the
+ * normalized range [-1.0, 1.0] (the native format of the WASAPI shared mixer).
+ * `out` must point to space for at least 2*frames float values.
  * Updates `*phase` in place. Returns the final gain value reached. */
-double synth_fill_buffer(int16_t* out, size_t frames,
+double synth_fill_buffer(float* out, size_t frames,
                          SynthPhase* phase,
                          SynthFrameParams p);
 
@@ -58,7 +59,7 @@ typedef struct {
 
 void noise_state_init(NoiseState* ns);
 
-double synth_fill_noise(int16_t* out, size_t frames, NoiseState* ns,
+double synth_fill_noise(float* out, size_t frames, NoiseState* ns,
                         int noise_type, double volume,
                         double gain_start, double gain_target, double gain_step);
